test: . Move last UI-related tests to tests/ui/QuickSearchTasksModal.test.ts
